add activity to l3 config
