+++ packages/config/src/common/index.ts
@@ -28,6 +28,7 @@ export * from './ScalingProjectStateDerivation'
 export * from './ScalingProjectStateValidation'
 export * from './ScalingProjectTechnology'
 export * from './ScalingProjectTechnologyChoice'
+export * from './ScalingProjectTransactionApi'
 export * from './stateCorrectness'
 export * from './stateDerivations'
 export * from './technology'
